+++ GodotEngine.Utility/Gizmos.cs
@@ -4,6 +4,13 @@ using System.Collections.Generic;
 namespace Cobilas.GodotEngine.Utility;
 /// <summary>Gizmos are used to give visual debugging or setup aids in the Scene view.</summary>
 public static class Gizmos {
+
+    /// <summary>Sets the Color of the gizmos that are drawn next.</summary>
+    /// <value>Returns or sets the color of the next gizmo.</value>
+    public static Color Color {
+        get => InternalGizmos.Color;
+        set => InternalGizmos.Color = value;
+    }
     /// <summary>
     /// Draws a line from a 2D point to another, with a given color and width. It can
     /// be optionally antialiased. See also <seealso cref="CanvasItem.DrawMultiline(Vector2[], Color, float, bool)"/>